* time_t is signed, not unsigned. Fixed.   -Should fix timesync, we'll see. * Added stuff to cmd_debug for figuring out if timesync is fixed. * Possibly fixed some cookie-ops issues
svn: 831

+++ src/mod/server.mod/server.c
@@ -1383,7 +1383,7 @@ static void server_5minutely()
       putlog(LOG_SERV, "*", IRC_SERVERSTONED);
     } else if (!trying_server) {
       /* Check for server being stoned. */
-      dprintf(DP_MODE, "PING :%lu\n", (unsigned long) now);
+      dprintf(DP_MODE, "PING :%li\n", now);
       waiting_for_awake = 1;
     }
   }
